+++ inst/climex_app/ui.R
@@ -1,1 +1,0 @@
-shinyUI( climex::climex.ui() )